+++ NEWS.md
@@ -1,3 +1,8 @@
+pool 0.1.5.9000
+================
+
+* Fixed #114: `dbPool()` objects previously could leak memory. (#115)
+
 pool 0.1.5
 ================